requirements.txt
# Основные зависимости
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
gunicorn==21.2.0

# Для работы с PostgreSQL на Render
psycopg2-binary==2.9.9
python-dotenv==1.0.0

# Дополнительные
Pillow==10.1.0
requests==2.31.0

# Для ускорения работы
blinker==1.7.0
greenlet==3.0.1
